extract image uploads into its own hook
